Boss - DefenderX
Added a boss spritesheet to the game. The boss has 3 blue orbs attached
to it. When 3 arrows hit any of the orbs on the wings then play the
corresponding animation which displays the ship without that wing. Only
when the orb at the very top gets hit 3 times will the ship be
destroyed.
I have all the animations to make this happen apart from the ship
exploding maybe. The ship should shoot a rocket at the ground when a ray
cast coming from the turret detects the player as it passes by. You
could increase the frequency of the rockets when the ship loses a wing,
could be fun. The could have basic AI which follows the player instead
of having set movement.